requirements.txt
Django==1.7.1
Markdown==2.5.2
django-admin-bootstrapped==2.3.1
djangorestframework==3.0.0
mysqlclient==1.3.4
